+++ Audit-Service/src/main/java/com/arivanamin/healthcare/backend/audit/infrastructure/AuditApplication.java
@@ -6,7 +6,7 @@ import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 
-import static com.arivanamin.healthcare.backend.core.domain.base.CoreApplicationConfig.BASE_PACKAGE;
+import static com.arivanamin.healthcare.backend.base.domain.base.CoreApplicationConfig.BASE_PACKAGE;
 
 @SpringBootApplication (scanBasePackages = BASE_PACKAGE)
 @EntityScan (BASE_PACKAGE)
