+++ Core/src/main/java/com/arivanamin/healthcare/backend/base/domain/aspects/LogExecutionTime.java
@@ -1,4 +1,4 @@
-package com.arivanamin.healthcare.backend.core.domain.aspects;
+package com.arivanamin.healthcare.backend.base.domain.aspects;
 
 import java.lang.annotation.*;
