+++ Core/src/main/java/com/arivanamin/healthcare/backend/base/domain/gender/Gender.java
@@ -1,0 +1,6 @@
+package com.arivanamin.healthcare.backend.base.domain.gender;
+
+public enum Gender {
+    
+    MALE, FEMALE, OTHER
+}
