--- Core/src/main/java/com/arivanamin/healthcare/backend/core/domain/gender/Gender.java
@@ -1,6 +1,0 @@
-package com.arivanamin.healthcare.backend.core.domain.gender;
-
-public enum Gender {
-    
-    MALE, FEMALE, OTHER
-}
