+++ Core/src/test/java/com/arivanamin/healthcare/backend/base/domain/base/CoreApplicationConfigTest.java
@@ -1,10 +1,10 @@
-package com.arivanamin.healthcare.backend.core.domain.base;
+package com.arivanamin.healthcare.backend.base.domain.base;
 
 import org.junit.jupiter.api.Test;
 
 import java.nio.file.Path;
 
-import static com.arivanamin.healthcare.backend.core.domain.base.CoreApplicationConfig.USER_HOME_DIRECTORY_PROPERTY;
+import static com.arivanamin.healthcare.backend.base.domain.base.CoreApplicationConfig.USER_HOME_DIRECTORY_PROPERTY;
 import static java.io.File.separator;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.springframework.test.util.ReflectionTestUtils.invokeMethod;
