+++ Patient-Service/src/test/java/com/arivanamin/healthcare/backend/patient/PatientArchitectureTest.java
@@ -5,7 +5,7 @@ import com.arivanamin.healthcare.backend.testing.architecture.rules.CommonBestPr
 import com.tngtech.archunit.core.importer.ImportOption;
 import com.tngtech.archunit.junit.AnalyzeClasses;
 
-import static com.arivanamin.healthcare.backend.core.domain.base.CoreApplicationConfig.BASE_PACKAGE;
+import static com.arivanamin.healthcare.backend.base.domain.base.CoreApplicationConfig.BASE_PACKAGE;
 
 @AnalyzeClasses (packages = BASE_PACKAGE, importOptions = ImportOption.DoNotIncludeTests.class)
 class PatientArchitectureTest implements CommonBestPracticeRules, CleanArchitectureRules {
