+++ Patient-Service/src/test/java/com/arivanamin/healthcare/backend/patient/PatientTestingBestPracticeRulesTest.java
@@ -4,7 +4,7 @@ import com.arivanamin.healthcare.backend.testing.architecture.rules.TestingBestP
 import com.tngtech.archunit.core.importer.ImportOption;
 import com.tngtech.archunit.junit.AnalyzeClasses;
 
-import static com.arivanamin.healthcare.backend.core.domain.base.CoreApplicationConfig.BASE_PACKAGE;
+import static com.arivanamin.healthcare.backend.base.domain.base.CoreApplicationConfig.BASE_PACKAGE;
 
 @AnalyzeClasses (packages = BASE_PACKAGE, importOptions = ImportOption.OnlyIncludeTests.class)
 class PatientTestingBestPracticeRulesTest implements TestingBestPracticeRules {
